+++ src/hooks/useDebouncedTrade.ts
@@ -1,0 +1,104 @@
+import { Currency, CurrencyAmount, TradeType } from '@uniswap/sdk-core'
+import { useWeb3React } from '@web3-react/core'
+import { WRAPPED_NATIVE_CURRENCY } from 'constants/tokens'
+import { DebounceSwapQuoteVariant, useDebounceSwapQuoteFlag } from 'featureFlags/flags/debounceSwapQuote'
+import { useMemo } from 'react'
+import { RouterPreference } from 'state/routing/types'
+import { ClassicTrade, InterfaceTrade, QuoteMethod, TradeState } from 'state/routing/types'
+import { useRoutingAPITrade } from 'state/routing/useRoutingAPITrade'
+import { useRouterPreference } from 'state/user/hooks'
+
+import useAutoRouterSupported from './useAutoRouterSupported'
+import useDebounce from './useDebounce'
+import useIsWindowVisible from './useIsWindowVisible'
+
+// Prevents excessive quote requests between keystrokes.
+const DEBOUNCE_TIME = 350
+
+// Temporary until we remove the feature flag.
+const DEBOUNCE_TIME_INCREASED = 650
+
+export function useDebouncedTrade(
+  tradeType: TradeType,
+  amountSpecified?: CurrencyAmount<Currency>,
+  otherCurrency?: Currency,
+  routerPreferenceOverride?: RouterPreference.X,
+  account?: string
+): {
+  state: TradeState
+  trade?: InterfaceTrade
+}
+
+export function useDebouncedTrade(
+  tradeType: TradeType,
+  amountSpecified?: CurrencyAmount<Currency>,
+  otherCurrency?: Currency,
+  routerPreferenceOverride?: RouterPreference.API | RouterPreference.CLIENT,
+  account?: string
+): {
+  state: TradeState
+  trade?: ClassicTrade
+}
+/**
+ * Returns the debounced v2+v3 trade for a desired swap.
+ * @param tradeType whether the swap is an exact in/out
+ * @param amountSpecified the exact amount to swap in/out
+ * @param otherCurrency the desired output/payment currency
+ * @param routerPreferenceOverride force useRoutingAPITrade to use a specific RouterPreference
+ * @param account the connected address
+ *
+ */
+export function useDebouncedTrade(
+  tradeType: TradeType,
+  amountSpecified?: CurrencyAmount<Currency>,
+  otherCurrency?: Currency,
+  routerPreferenceOverride?: RouterPreference,
+  account?: string
+): {
+  state: TradeState
+  trade?: InterfaceTrade
+  method?: QuoteMethod
+} {
+  const { chainId } = useWeb3React()
+  const autoRouterSupported = useAutoRouterSupported()
+  const isWindowVisible = useIsWindowVisible()
+
+  const debouncedSwapQuoteFlagEnabled = useDebounceSwapQuoteFlag() === DebounceSwapQuoteVariant.Enabled
+  const [debouncedAmount, debouncedOtherCurrency] = useDebounce(
+    useMemo(() => [amountSpecified, otherCurrency], [amountSpecified, otherCurrency]),
+    debouncedSwapQuoteFlagEnabled ? DEBOUNCE_TIME_INCREASED : DEBOUNCE_TIME
+  )
+
+  const isAWrapTransaction = useMemo(() => {
+    if (!chainId || !amountSpecified || !debouncedOtherCurrency) return false
+    const weth = WRAPPED_NATIVE_CURRENCY[chainId]
+    return (
+      (amountSpecified.currency.isNative && weth?.equals(debouncedOtherCurrency)) ||
+      (debouncedOtherCurrency.isNative && weth?.equals(amountSpecified.currency))
+    )
+  }, [amountSpecified, chainId, debouncedOtherCurrency])
+
+  const shouldGetTrade = !isAWrapTransaction && isWindowVisible
+
+  const [routerPreference] = useRouterPreference()
+  const routingAPITrade = useRoutingAPITrade(
+    tradeType,
+    amountSpecified ? debouncedAmount : undefined,
+    debouncedOtherCurrency,
+    routerPreferenceOverride ?? routerPreference,
+    !(autoRouterSupported && shouldGetTrade), // skip fetching
+    account
+  )
+
+  const inDebounce =
+    (!debouncedAmount && Boolean(amountSpecified)) || (!debouncedOtherCurrency && Boolean(otherCurrency))
+  const isLoading = routingAPITrade.state === TradeState.LOADING || inDebounce
+
+  return useMemo(
+    () => ({
+      ...routingAPITrade,
+      ...(isLoading ? { state: TradeState.LOADING } : {}),
+    }),
+    [isLoading, routingAPITrade]
+  )
+}
