+++ src/nft/hooks/useDerivedPayWithAnyTokenSwapInfo.ts
@@ -1,6 +1,6 @@
 import { Currency, CurrencyAmount, NativeCurrency, Percent, Token, TradeType } from '@uniswap/sdk-core'
 import useAutoSlippageTolerance from 'hooks/useAutoSlippageTolerance'
-import { useBestTrade } from 'hooks/useBestTrade'
+import { useDebouncedTrade } from 'hooks/useDebouncedTrade'
 import { useMemo } from 'react'
 import { RouterPreference } from 'state/routing/types'
 import { ClassicTrade, TradeState } from 'state/routing/types'
@@ -15,7 +15,7 @@ export default function useDerivedPayWithAnyTokenSwapInfo(
   maximumAmountIn?: CurrencyAmount<Token>
   allowedSlippage: Percent
 } {
-  const { state, trade } = useBestTrade(
+  const { state, trade } = useDebouncedTrade(
     TradeType.EXACT_OUTPUT,
     parsedOutputAmount,
     inputCurrency ?? undefined,
